+++ internal/models/order.go
@@ -3,22 +3,22 @@ package models
 import "time"
 
 type Order struct {
-	ID          uint64    `gorm:"primary_key"`
-	UserID      uint64    `gorm:"not null"`
+	ID          uint      `gorm:"primary_key"`
+	UserID      uint      `gorm:"not null"`
 	OrderNumber string    `gorm:"unique;not null"`
 	Status      string    `gorm:"default:'NEW'"`
-	Accrual     float32   `gorm:"default:null"`
-	Sum         float32   `gorm:"default:null"`
+	Accrual     float64   `gorm:"default:null"`
+	Sum         float64   `gorm:"default:null"`
 	UploadedAt  time.Time `gorm:"autoCreateTime"`
 	UpdatedAt   time.Time `gorm:"autoUpdateTime"`
 }
 
 type OrdersUserJSON struct {
-	ID          uint64    `gorm:"primary_key" json:"-"`
-	UserID      uint64    `gorm:"not null" json:"-"`
+	ID          uint      `gorm:"primary_key" json:"-"`
+	UserID      uint      `gorm:"not null" json:"-"`
 	OrderNumber string    `gorm:"unique;not null" json:"number"`
 	Status      string    `gorm:"default:'NEW'" json:"status"`
-	Accrual     float32   `gorm:"default:null" json:"accrual,omitempty"`
+	Accrual     float64   `gorm:"default:null" json:"accrual,omitempty"`
 	UploadedAt  time.Time `gorm:"autoCreateTime" json:"uploaded_at"`
 	UpdatedAt   time.Time `gorm:"autoUpdateTime" json:"-"`
 }
@@ -27,5 +27,5 @@ type OrdersUserJSON struct {
 type AccrualInfo struct {
 	OrderNumber string `json:"order"`
 	Status      string
-	Accrual     float32 `json:"accrual"`
+	Accrual     float64 `json:"accrual"`
 }
